Implement All Instructions
This commit includes implementing all instructions except for (BRK and
RTI). It also includes a large refactor of instructions unifying
addressing code and instruction code rather than duplicating it for each
permutation.

There are certainly more ways that this code can be reduced and shared,
but this is a good attempt at some order.